Add 2fa optoin on settings
