+++ frontend/containers/open-call-page/funding-impact/component.tsx
@@ -110,7 +110,7 @@ export const OpenCallFundingImpact: FC<OpenCallFundingImpactProps> = ({
                 <h3 className="mb-6 text-xl font-semibold">
                   <FormattedMessage defaultMessage="SDG's" id="d3TPmn" />
                 </h3>
-                <SDGs sdgs={openCallSdgs} />
+                <SDGs size="large" sdgs={openCallSdgs} />
               </div>
             </div>
           </LayoutContainer>
