+++ frontend/containers/settings-security/2fa/index.ts
@@ -1,0 +1,2 @@
+export { default } from './component';
+export type { Settings2FAProps } from './types';
