+++ frontend/containers/settings-security/2fa/types.ts
@@ -1,0 +1,1 @@
+export type Settings2FAProps = {};
