+++ frontend/containers/settings-security/change-password/index.ts
@@ -1,0 +1,2 @@
+export { default } from './component';
+export type { SettingsChangePasswordProps } from './types';
