+++ frontend/containers/settings-security/change-password/types.ts
@@ -1,0 +1,1 @@
+export type SettingsChangePasswordProps = {};
